requirements.txt
torch>=2.0.0
Pillow>=9.0.0
numpy>=1.22.0
opencv-python-headless>=4.5.0
requests>=2.26.0
transformers>=4.30.0
comfyui-extra-nodes>=1.4.0